// src/app/core/enums/type.enum.ts
export enum Type {
  LOGO_CREATION = 'LOGO_CREATION',
  BRAND_CREATION = 'BRAND_CREATION',
  UIUX = 'UIUX',
  MARKETING = 'MARKETING',
  DESIGN = 'DESIGN',
  WEB_DEVELOPMENT = 'WEB_DEVELOPMENT',
  DESKTOP_DEVELOPMENT = 'DESKTOP_DEVELOPMENT',
  MOBILE_APP_DEVELOPMENT = 'MOBILE_APP_DEVELOPMENT',
  AUTOMATION = 'AUTOMATION',
  TECHNICAL_SUPPORT = 'TECHNICAL_SUPPORT',
  OTHER = 'OTHER',
}